update portal to home
